Move useKpiData hook to appropriate location
Also, do not export KpiView anymore.

JIRA: RAIL-2956
