+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/useKpiData.ts
@@ -27,8 +27,8 @@ import {
     useWorkspace,
 } from "@gooddata/sdk-ui";
 import invariant from "ts-invariant";
-import { filterContextItemsToFiltersForWidget, filterContextToFiltersForWidget } from "../../converters";
-import { IDashboardFilter } from "../../types";
+import { filterContextItemsToFiltersForWidget, filterContextToFiltersForWidget } from "../converters";
+import { IDashboardFilter } from "../types";
 
 interface IUseKpiDataConfig {
     kpiWidget?: IKpiWidget;
@@ -46,6 +46,9 @@ interface IUseKpiDataResult {
     allFilters?: IDashboardFilter[];
 }
 
+/**
+ * @internal
+ */
 export function useKpiData({
     kpiWidget,
     filters,
